requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
diskcache==5.6.3
emoji==2.10.1
environs==9.5.0
grpcio==1.60.0
idna==3.6
Jinja2==3.1.3
llama_cpp_python==0.2.56
MarkupSafe==2.1.5
marshmallow==3.21.1
minio==7.2.5
numpy==1.26.4
packaging==24.0
pandas==2.2.1
protobuf==4.25.3
pyarrow==15.0.1
pycparser==2.21
pycryptodome==3.20.0
pymilvus==2.3.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
